app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
azure-storage-blob==12.24.0
azure-identity==1.19.0
pydantic==2.11.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
httpx==0.28.0
